v5.6 (fix) | Component file restructure
