+++ src/lib/index.js
@@ -11,14 +11,14 @@ export { default as LoginLink } from './inputs/Login-link.svelte'
 
 // Dashboard / Navigation
 export { default as Pagination } from './navigation/Pagination.svelte'
+export { default as Toolbar } from './navigation/Toolbar.svelte'
 
 // Dashboard / inputs
 export { default as PrimaryButton } from './inputs/Primary-button.svelte'
 export { default as Select } from './inputs/Select.svelte'
-export { default as Toolbar } from './navigation/Toolbar.svelte'
-export { default as Search } from './navigation/Search.svelte'
-export { default as Sort } from './navigation/Sort.svelte'
-export { default as Filter } from './navigation/Filter.svelte'
+export { default as Search } from './inputs/Search.svelte'
+export { default as Sort } from './inputs/Sort.svelte'
+export { default as Filter } from './inputs/Filter.svelte'
 
 export { default as Language } from './inputs/Language.svelte'
 // export { default as Component } from './path/Component.svelte'
